N4 — the nnsight tracing engine: module boundaries, proven by parity
The second implementation of the engine contract, built on nnsight 0.8
pinned at the rev the N0 run verified on the real 35B. This skeleton serves
the module-boundary vocabulary; the interiors it exists for (DeltaNet
state, expert interiors, attention internals) arrive N5-N8.

What made it small: the component→module map resolves against nnsight's
envoy tree unchanged, and the executor's document half was never
hook-specific. So this PR finishes the shared extraction N3 started —
sites.py, featurizers.py, the execute() orchestration (shared/execution.py)
and the executor core (shared/executor_base.py) move to neural/shared/ —
and the two engines become thin: PointExecutor is the hook half
(_run_group, decode, install/capture plumbing), TracePointExecutor the
trace half (~150 lines: one trace per group, writes assign envoy
input/output, reads nnsight.save the contract tensor, operations issued in
forward order).

The parity suite is the correctness proof and the handoff's numerical
oracle in one: 27 read cases and 12 interchange-write cases across both
fixture families (tiny-llama + the hybrid tiny qwen3.5-moe) agree at 1e-5
in fp32/eager/CPU; policy refusals are string-identical because the tables
are single-homed; attention_probs routes to the reference engine.

Not declared, on purpose: grad (D6), generate (N8), attention_probs /
writable_attention_probs (N5), quantized_weights (unverified). Documents
needing them route to pytorch_hooks under --engine auto.

Deps: 'nnsight' optional extra + dev group, one pinned git source
(tool.uv.sources, rev 8c480727); CI installs the dev group so the parity
suite is a real gate, and the tests importorskip for envs without it.

Gates: 2005 passed (-m 'not golden'), corpus digests zero diff, pre-commit
clean.

Plan: cockpit notes/causalab-nnsight-engine-plan.md §4 / §6-N4 (D1, D8).
